requirements.txt
streamlit
pandas
numpy
plotly
requests
beautifulsoup4
lxml
selenium
webdriver-manager
python-dotenv
SQLAlchemy